All non-404 errors are formatted as HandlerErrors
