+++ auth/auth.go
@@ -30,7 +30,7 @@ import (
     "github.com/gorilla/mux"
 
     "github.com/lighthouse/lighthouse/users"
-
+    "github.com/lighthouse/lighthouse/handlers"
     "github.com/lighthouse/lighthouse/session"
 )
 
@@ -96,7 +96,7 @@ func AuthMiddleware(h http.Handler, ignorePaths []string) http.Handler {
         }
 
         if strings.HasPrefix(r.URL.Path, "/api") {
-            w.WriteHeader(401)
+            handlers.WriteError(w, 401, "auth", "not logged in")
         } else {
             http.Redirect(w, r, "/", http.StatusFound)
         }
@@ -141,7 +141,7 @@ func Handle(r *mux.Router) {
         if userOK && passwordOK {
             w.WriteHeader(200)
         } else {
-            w.WriteHeader(401)
+            handlers.WriteError(w, 401, "auth", "email or password incorrect")
         }
 
         fmt.Fprintf(w, "%t", userOK && passwordOK)
