+++ beacons/handlers.go
@@ -16,9 +16,7 @@ package beacons
 
 import (
 	"fmt"
-    "time"
     "strconv"
-    "net"
 	"net/http"
 	"encoding/json"
 	"io/ioutil"
@@ -40,25 +38,17 @@ func getInstanceAlias(instance string) string {
 }
 
 func writeResponse(err error, w http.ResponseWriter) {
-    var code int
-
     switch err {
+        case nil:
+            w.WriteHeader(http.StatusOK)
+
         case databases.KeyNotFoundError, databases.NoUpdateError, 
                 databases.EmptyKeyError, NotEnoughParametersError:
-            code = http.StatusBadRequest
-
-        case nil:
-            code = http.StatusOK
+            handlers.WriteError(w, http.StatusBadRequest, "beacons", err.Error())
 
         default:
-            code = http.StatusInternalServerError
-    }
-
-    w.WriteHeader(code)
-
-    if err != nil {
-        fmt.Fprint(w, err)
-    }
+            handlers.WriteError(w, http.StatusInternalServerError, "beacons", err.Error())
+    }    
 }
 
 func handleAddUserToBeacon(w http.ResponseWriter, r *http.Request) {
@@ -86,9 +76,12 @@ func handleAddUserToBeacon(w http.ResponseWriter, r *http.Request) {
 }
 
 func handleRemoveUserFromBeacon(w http.ResponseWriter, r *http.Request) {
+    var err error = nil
+    defer func() { writeResponse(err, w) }()
+
     params, ok := handlers.GetEndpointParams(r, []string{"Beacon", "UserId"})
     if ok == false || len(params) < 2 {
-        writeResponse(NotEnoughParametersError, w)
+        err = NotEnoughParametersError
         return
     }
 
@@ -101,13 +94,16 @@ func handleRemoveUserFromBeacon(w http.ResponseWriter, r *http.Request) {
         err = updateBeaconField("Users", data.Users, beacon)
     }
 
-    writeResponse(err, w)
+    return
 }
 
 func handleUpdateBeaconToken(w http.ResponseWriter, r *http.Request) {
+    var err error = nil
+    defer func() { writeResponse(err, w) }()
+
     params, ok := handlers.GetEndpointParams(r, []string{"Beacon"})
     if ok == false || len(params) < 1 {
-        writeResponse(NotEnoughParametersError, w)
+        err = NotEnoughParametersError
         return
     }
 
@@ -115,7 +111,6 @@ func handleUpdateBeaconToken(w http.ResponseWriter, r *http.Request) {
 
     reqBody, err := ioutil.ReadAll(r.Body)
     if err != nil {
-        writeResponse(err, w)
         return
     }
 
@@ -126,7 +121,7 @@ func handleUpdateBeaconToken(w http.ResponseWriter, r *http.Request) {
         err = updateBeaconField("Token", token, beacon)
     }
 
-    writeResponse(err, w)
+    return
 }
 
 func handleBeaconCreate(w http.ResponseWriter, r *http.Request) {
@@ -150,8 +145,6 @@ func handleBeaconCreate(w http.ResponseWriter, r *http.Request) {
         return
     }
 
-    fmt.Println(beaconInfo.Address, beaconInfo.Alias)
-
     if beaconInfo.Address == "" || beaconInfo.Alias == "" {
         err = NotEnoughParametersError
         return
@@ -166,11 +159,11 @@ func handleBeaconCreate(w http.ResponseWriter, r *http.Request) {
         beacon.Users[user] = true
     }
 
-    _, err = net.DialTimeout("ip", "http://" + beacon.Address, 
-        time.Duration(3) * time.Second)
-    if err != nil {
-        //return
-    }
+    // _, err = net.DialTimeout("ip", "http://" + beacon.Address, 
+    //     time.Duration(3) * time.Second)
+    // if err != nil {
+    //     return
+    // }
 
     err = aliases.AddAlias(beaconInfo.Alias, beaconInfo.Address)
     if err != nil {
