+++ handlers/handlers.go
@@ -23,8 +23,6 @@ import (
     "io/ioutil"
 
     "github.com/gorilla/mux"
-
-    "github.com/lighthouse/lighthouse/beacons/aliases"
 )
 
 /*
@@ -68,7 +66,7 @@ type CustomHandlerMap map[*regexp.Regexp]CustomHandlerFunc
     Container of failure data created by the handlers
 */
 type HandlerError struct {
-    StatusCode  int
+    StatusCode  int     `json:"-"`
     Cause       string
     Message     string
 }
@@ -129,7 +127,7 @@ func Rollback(
     info HandlerInfo,
     runHandlers []CustomHandlerFunc,
 ) {
-    WriteError(w, err)
+    WriteError(w, err.StatusCode, err.Cause, err.Message)
     for _, handler := range runHandlers {
         handler(info, true)
     }
@@ -138,46 +136,13 @@ func Rollback(
 /*
     Writes error data and code to the HTTP response.
 */
-func WriteError(w http.ResponseWriter, err HandlerError) {
-    json, _ := json.Marshal(struct {
-        Error   string
-        Message string
-    }{err.Cause, err.Message})
-
+func WriteError(w http.ResponseWriter, code int, cause, message string) {
     w.Header().Set("Content-Type", "application/json")
-    w.WriteHeader(err.StatusCode)
+    w.WriteHeader(code)
+
+    err := HandlerError{Cause: cause, Message: message}
+    json, _ := json.Marshal(err)
     w.Write(json)
-}
-
-/*
-    Extracts data from the request to create a HandlerInfo
-    which is used by the handlers.
-
-    RETURN: A HandlerInfo extracted from the request
-*/
-func GetHandlerInfo(r *http.Request) (HandlerInfo, bool) {
-    var info HandlerInfo
-
-    params, ok := GetEndpointParams(r, []string{"Host", "DockerEndpoint"})
-
-    if ok == false || len(params) < 2 {
-        return HandlerInfo{}, false
-    }
-
-    hostAlias := params["Host"]
-
-    value, err := aliases.GetAddressOf(hostAlias)
-    if err == nil {
-        info.Host = value
-    } else {
-        info.Host = hostAlias // Unknown alias, just use what was given
-    }
-
-    info.DockerEndpoint = params["DockerEndpoint"]
-    info.Body = GetRequestBody(r)
-    info.Request = r
-
-    return info, true
 }
 
 /*
